requirements.txt
fal-client
python-dotenv
elevenlabs
streamlit
st-clickable-images
web-browser

